requirements-dev.txt
appdirs==1.4.4
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
Django==3.2.7
django-advanced-filters==1.3.0
django-braces==1.14.0
django-cors-headers==3.8.0
django-filter==2.4.0
django-reversion==3.0.9
djangorestframework==3.12.4
djangorestframework-gis==0.17
djangorestframework-simplejwt==4.8.0
filelock==3.0.12
platformdirs==2.3.0
PyJWT==2.1.0
pytz==2021.1
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
virtualenv==20.8.0
